requirements.txt
minswap-py=0.3.4
pycardano=0.9.0
solara=1.19.0